+++ packages/camera/lib/camera.dart
@@ -58,6 +58,7 @@ Future<List<CameraDescription>> availableCameras() async {
       return CameraDescription(
         name: camera['name'],
         lensDirection: _parseCameraLensDirection(camera['lensFacing']),
+        sensorOrientation: camera['sensorOrientation'],
       );
     }).toList();
   } on PlatformException catch (e) {
@@ -66,10 +67,19 @@ Future<List<CameraDescription>> availableCameras() async {
 }
 
 class CameraDescription {
-  CameraDescription({this.name, this.lensDirection});
+  CameraDescription({this.name, this.lensDirection, this.sensorOrientation});
 
   final String name;
   final CameraLensDirection lensDirection;
+
+  /// Clockwise angle through which the output image needs to be rotated to be upright on the device screen in its native orientation.
+  ///
+  /// **Range of valid values:**
+  /// 0, 90, 180, 270
+  ///
+  /// On Android, also defines the direction of rolling shutter readout, which
+  /// is from top to bottom in the sensor's coordinate system.
+  final int sensorOrientation;
 
   @override
   bool operator ==(Object o) {
@@ -85,7 +95,7 @@ class CameraDescription {
 
   @override
   String toString() {
-    return '$runtimeType($name, $lensDirection)';
+    return '$runtimeType($name, $lensDirection, $sensorOrientation)';
   }
 }