+++ packages/in_app_purchase/lib/src/store_kit_wrappers/sk_payment_queue_wrapper.dart
@@ -3,6 +3,7 @@
 // found in the LICENSE file.
 
 import 'dart:async';
+import 'package:collection/collection.dart';
 import 'package:flutter/foundation.dart';
 import 'package:in_app_purchase/src/channel.dart';
 import 'package:json_annotation/json_annotation.dart';
@@ -214,7 +215,7 @@ enum SKPaymentTransactionStateWrapper {
   /// Indicates the transaction is being processed in App Store.
   ///
   /// You should update your UI to indicate the process and waiting for the transaction to update to the next state.
-  /// Never complte a transaction that is in purchasing state.
+  /// Never complete a transaction that is in purchasing state.
   @JsonValue(0)
   purchasing,
 
@@ -302,6 +303,24 @@ class SKPaymentTransactionWrapper {
 
   /// The error object, only available if the [transactionState] is [SKPaymentTransactionStateWrapper.failed].
   final SKError error;
+
+  @override
+  bool operator ==(Object other) {
+    if (identical(other, this)) {
+      return true;
+    }
+    if (other.runtimeType != runtimeType) {
+      return false;
+    }
+    final SKPaymentTransactionWrapper typedOther = other;
+    return typedOther.payment == payment &&
+        typedOther.transactionState == transactionState &&
+        typedOther.originalTransaction == originalTransaction &&
+        typedOther.transactionTimeStamp == transactionTimeStamp &&
+        typedOther.transactionIdentifier == transactionIdentifier &&
+        DeepCollectionEquality().equals(typedOther.downloads, downloads) &&
+        typedOther.error == error;
+  }
 }
 
 /// Dart wrapper around StoreKit's [SKDownloadState](https://developer.apple.com/documentation/storekit/skdownloadstate?language=objc).
@@ -403,6 +422,27 @@ class SKDownloadWrapper {
 
   /// The error that prevented the downloading; only available if the [transactionState] is [SKPaymentTransactionStateWrapper.failed].
   final SKError error;
+
+  @override
+  bool operator ==(Object other) {
+    if (identical(other, this)) {
+      return true;
+    }
+    if (other.runtimeType != runtimeType) {
+      return false;
+    }
+    final SKDownloadWrapper typedOther = other;
+    return typedOther.contentIdentifier == contentIdentifier &&
+        typedOther.state == state &&
+        typedOther.contentLength == contentLength &&
+        typedOther.contentURL == contentURL &&
+        typedOther.contentVersion == contentVersion &&
+        typedOther.transactionID == transactionID &&
+        typedOther.progress == progress &&
+        typedOther.timeRemaining == timeRemaining &&
+        typedOther.downloadTimeUnknown == downloadTimeUnknown &&
+        typedOther.error == error;
+  }
 }
 
 /// Dart wrapper around StoreKit's [NSError](https://developer.apple.com/documentation/foundation/nserror?language=objc).
@@ -430,6 +470,21 @@ class SKError {
 
   /// A map that contains more detailed information about the error. Any key of the map must be one of the [NSErrorUserInfoKey](https://developer.apple.com/documentation/foundation/nserroruserinfokey?language=objc).
   final Map<String, dynamic> userInfo;
+
+  @override
+  bool operator ==(Object other) {
+    if (identical(other, this)) {
+      return true;
+    }
+    if (other.runtimeType != runtimeType) {
+      return false;
+    }
+    final SKError typedOther = other;
+    return typedOther.code == code &&
+        typedOther.domain == domain &&
+        DeepCollectionEquality.unordered()
+            .equals(typedOther.userInfo, userInfo);
+  }
 }
 
 /// Dart wrapper around StoreKit's [SKPayment](https://developer.apple.com/documentation/storekit/skpayment?language=objc).
@@ -474,7 +529,7 @@ class SKPaymentWrapper {
   /// An opaque id for the user's account.
   ///
   /// Used to help the store detect irregular activity. See https://developer.apple.com/documentation/storekit/skpayment/1506116-applicationusername?language=objc for more details.
-  /// For example, you can use a one-way hash of the user’s account name on your server. Don’t use the Apple ID for your developer account, the user’s Apple ID, or the user’s unhashed account name on your server.
+  /// For example, you can use a one-way hash of the user’s account name on your server. Don’t use the Apple ID for your developer account, the user’s Apple ID, or the user’s not hashed account name on your server.
   final String applicationUsername;
 
   /// Reserved for future use.
@@ -493,4 +548,20 @@ class SKPaymentWrapper {
   ///
   /// For how to test in App Store sand box, see https://developer.apple.com/in-app-purchase/.
   final bool simulatesAskToBuyInSandbox;
+
+  @override
+  bool operator ==(Object other) {
+    if (identical(other, this)) {
+      return true;
+    }
+    if (other.runtimeType != runtimeType) {
+      return false;
+    }
+    final SKPaymentWrapper typedOther = other;
+    return typedOther.productIdentifier == productIdentifier &&
+        typedOther.applicationUsername == applicationUsername &&
+        typedOther.quantity == quantity &&
+        typedOther.simulatesAskToBuyInSandbox == simulatesAskToBuyInSandbox &&
+        typedOther.requestData == requestData;
+  }
 }